BUGME, one more qt5 followup, make it compile with Qt switchable environment, and packaging from http://ppa.launchpad.net/canonical-qt5-edgers/qt5-proper/ubuntu/pool/main/q/
Use qmake for runtime version detection, use strict swith between Qt versions without mixing includes and libs

+++ gfx/thebes/gfxQtPlatform.cpp
@@ -203,7 +203,7 @@ gfxQtPlatform::GetXScreen(QWidget* aWindow)
 #endif
 #else
   return ScreenOfDisplay(GetXDisplay(aWindow),
-                         (int)qApp->platformNativeInterface()->
+                         (int)(intptr_t)qApp->platformNativeInterface()->
                            nativeResourceForWindow("screen",
                              aWindow ? aWindow->windowHandle() : nullptr));
 #endif
